Redefine temporal depth, explanation depth, and plan length for progression node.
The concept of temporal depth has now been defined in such a way to better allow for breadth-first-search-like behavior. The concept of explanation depth is now similar to what temporal offset + temporal depth was before. The concept of plan length has also been redefined to include the action leading to the root plan. The concept of temporal offset has been removed.