Solution Improvements
Add methods to the state graph node and state graph temporal edge classes to determine whether a solution is valid. Modify solution slightly so that an explanation can be gotten by index. Add a method to ProgressionTreeSpace for getting the plan and full history of actions represented by a node ID.